zstd: Fix incorrect pragma error on LCC compiler
LCC (Elbrus C Compiler) doesn't understand some of
GCC pragmas, despite of declaring itself GCC-aware.
The pragma of subject is the one that forbids vectorizing.
Actually, LCC don't vectorize anything unless explicitly
said to, so this pragma may be safely omitted and thus
not cause an error. This patch does this.