Help: Clarify valid versus supported values for CUDA_STANDARD
CMake 3.18 added the first support for any compiler for 17 and 20,
but those were recognized as valid values in earlier CMake versions
even though there was no compiler that supported them. Make this
distinction clear to avoid creating the impression that these standards
could be usefully used before CMake 3.18.

While 98 is recognized as a valid value, it also just gets treated as 03
internally. Document this behavior as well.

Fixes: #22711